driver/requirements.txt
pyserial
python-rtmidi